+++ config/config_test.go
@@ -24,8 +24,9 @@ func setupViper(t *testing.T) {
 	viper.SetConfigName(strings.ToLower(system.GetApplicationName()))
 	viper.AddConfigPath("/tmp/")
 	viper.Set(openai_key, "test_key")
-	viper.Set(openai_temperature, 0.2)
 	viper.Set(openai_proxy, "test_proxy")
+	viper.Set(openai_temperature, 0.2)
+	viper.Set(openai_max_tokens, 2000)
 	viper.Set(user_default_prompt_mode, "exec")
 	viper.Set(user_preferences, "test_preferences")
 
@@ -47,6 +48,7 @@ func testNewConfig(t *testing.T) {
 	assert.Equal(t, "test_key", cfg.GetAiConfig().GetKey())
 	assert.Equal(t, "test_proxy", cfg.GetAiConfig().GetProxy())
 	assert.Equal(t, 0.2, cfg.GetAiConfig().GetTemperature())
+	assert.Equal(t, 2000, cfg.GetAiConfig().GetMaxTokens())
 	assert.Equal(t, "exec", cfg.GetUserConfig().GetDefaultPromptMode())
 	assert.Equal(t, "test_preferences", cfg.GetUserConfig().GetPreferences())
 
@@ -61,16 +63,18 @@ func testWriteConfig(t *testing.T) {
 	require.NoError(t, err)
 
 	assert.Equal(t, "new_test_key", cfg.GetAiConfig().GetKey())
-	assert.Equal(t, 0.2, cfg.GetAiConfig().GetTemperature())
 	assert.Equal(t, "test_proxy", cfg.GetAiConfig().GetProxy())
+	assert.Equal(t, 0.2, cfg.GetAiConfig().GetTemperature())
+	assert.Equal(t, 2000, cfg.GetAiConfig().GetMaxTokens())
 	assert.Equal(t, "exec", cfg.GetUserConfig().GetDefaultPromptMode())
 	assert.Equal(t, "test_preferences", cfg.GetUserConfig().GetPreferences())
 
 	assert.NotNil(t, cfg.GetSystemConfig())
 
 	assert.Equal(t, "new_test_key", viper.GetString(openai_key))
-	assert.Equal(t, 0.2, viper.GetFloat64(openai_temperature))
 	assert.Equal(t, "test_proxy", viper.GetString(openai_proxy))
+	assert.Equal(t, 0.2, viper.GetFloat64(openai_temperature))
+	assert.Equal(t, 2000, viper.GetInt(openai_max_tokens))
 	assert.Equal(t, "exec", viper.GetString(user_default_prompt_mode))
 	assert.Equal(t, "test_preferences", viper.GetString(user_preferences))
 }